Add encrypt and decrypt
The current encrypt and decrypt may get changed because they use the binary tree that is created from the letter/frequency input.
Moving forward, I am either going to make a method to create a tree based on the letter/binary input or
just make static encrypt and decrypt methods that take a file.