Fix the output method and add english files
English files are the listed frequencies of the enlish alphabet and the binary for each letter.
I had to make it so instead of accounting for the variable i to increment till it reached the size of the priority queue,
the priority queue needed to slowly decrease and once the size reached 0 then the for loop stopped.